+++ src/qua_dashboards/video_mode/__init__.py
@@ -1,6 +1,6 @@
 from qua_dashboards.video_mode.dash_tools import *
 from qua_dashboards.video_mode.sweep_axis import SweepAxis
-from qua_dashboards.video_mode.voltage_parameters import *
+from qua_dashboards.video_mode.voltage_parameter import *
 from qua_dashboards.video_mode.inner_loop_actions import *
 from qua_dashboards.video_mode.scan_modes import *
 from qua_dashboards.video_mode.data_acquirers import *
