+++ src/qua_dashboards/video_mode/data_acquirers/base_data_aqcuirer.py
@@ -6,10 +6,10 @@ import numpy as np
 from dash import html
 import dash_bootstrap_components as dbc
 
+from qua_dashboards.utils.dash_utils import create_input_field
 from qua_dashboards.video_mode.sweep_axis import SweepAxis
 from qua_dashboards.video_mode.dash_tools import (
     create_axis_layout,
-    create_input_field,
     BaseDashComponent,
     ModifiedFlags,
 )
