+++ src/qua_dashboards/video_mode/sweep_axis.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 import numpy as np
 
-from qua_dashboards.video_mode.voltage_parameters import VoltageParameter
+from qua_dashboards.video_mode.voltage_parameter import VoltageParameter
 
 
 __all__ = ["SweepAxis"]
